Add the game search activity (#20)
This is pretty much the minimum requirement to meet the issue criteria.
There is no list of games, and no ability to actually add them. But,
that stuff seems pretty well covered by #17, so I'm fine with leaving it
for now.

This adds the game search activity. You can press the plus on the main
activity to get there. It has a search bar and search button that do
nothing (well, they show a "not implemented" prompt), and a back button
to go back to the watchlist. It also still has the settings button.

For the moment I have duplicated the top bar, but we should probably
find a way to avoid that. Apparently fragments might be the way?

I have switched the add button to the left, so that it can look better
when replaced with a back button.

I don't feel _great_ about all the meta file changes that this includes.
I recently updated android studio (through much pain and an eventual
reinstallation), so maybe that has something to do with it. _Apparently_
you are meant to commit these files, so I have left them in, but if we
end up in a battle between two android studios switching stuff back and
forth we may have to rethink this.

Closes #8.