Add database basics, new types, and new adapter
Way too much for one commit but it was all kind of entangled. I have
added an interface that all "game" objects we deal with must implement,
motivated by eventually wanting to use the same list adapter for the
game adding page and the watchlist page.

There are two implementations of this interface, the WatchlistGame,
which is the old one, and now also has some database shenanigans, and
SearchResultGame, which is just WatchlistGame but without a threshold,
and is designed to eventually be the one that appears on add game search
activity thingo.

I have also made a new adapter, because the recycler view list adapter
looks like it plays really nice with the LiveData stuff that you can do
with the database (to essentially have an list that automatically
updates with the database).

Oh and all the initial stuff required for the database should be here
too. Including a view model which is as of yet unused but wraps the
database stuff.