Make the add game page add to the database
This also involed sending the game object back through an intent, as it
is actually the watchlist page that does the adding, which wasn't
trivial. Only as I write this an I realising that the add page maybe
could have just directly used the database... oh well, it's done now, I
don't think this is much worse.

The watchlist still just displays the hardcoded list, but changing that
comes next.