Local database and adding games (#27)
Closes #16.
Closes #17.

* Add database basics, new types, and new adapter

Way too much for one commit but it was all kind of entangled. I have
added an interface that all "game" objects we deal with must implement,
motivated by eventually wanting to use the same list adapter for the
game adding page and the watchlist page.

There are two implementations of this interface, the WatchlistGame,
which is the old one, and now also has some database shenanigans, and
SearchResultGame, which is just WatchlistGame but without a threshold,
and is designed to eventually be the one that appears on add game search
activity thingo.

I have also made a new adapter, because the recycler view list adapter
looks like it plays really nice with the LiveData stuff that you can do
with the database (to essentially have an list that automatically
updates with the database).

Oh and all the initial stuff required for the database should be here
too. Including a view model which is as of yet unused but wraps the
database stuff.

* Improve readme

Doesn't really fit here, but I did it because I wanted to add something
else that I didn't end up adding, and I didn't want to go through the
process of making that change on another branch.

* Make the add game page add to the database

This also involed sending the game object back through an intent, as it
is actually the watchlist page that does the adding, which wasn't
trivial. Only as I write this an I realising that the add page maybe
could have just directly used the database... oh well, it's done now, I
don't think this is much worse.

The watchlist still just displays the hardcoded list, but changing that
comes next.

* Display the database watchlist

After all that setup work this was a very quick change.