Setup linting
As always this was more of a hassle than it feels like it should be. I
have added details describing most of what is going on here to the
firebase folder readme.

This also includes the changes to make firestore-tests pass the linting
(just following the config rules from firebase). Doesn't include
functions changes yet because I am in the process of actually writing
those functions.