Add handling of user and game watchlist syncing
I now have a firebase function that updates the watching users
collection on a game whenever a users watchlist item gets created
corresponding to that game.

I had to be a little careful because it is very easy to create infinite
loops here. I used zod out of habit, to ensure that all the unknown
objects I was injesting were okay, I hope that's alright.