Firestore schema and initial syncing work (#33)
* Run firebase init

I've set up stuff (including emulators) for functions, firestore, and
auth.

* Setup the android app to use the emulators

If the build is a debug build, as the usual emulator builds are. You can
also set a flag to force it to use production, if you want.

EDIT: I changed this in a merge to default to not using the emulators.

* Add rules for user watchlists and related tests

Hopefully this is all of the firestore rules I need for this PR, but
this can definitely be expanded (say, for device stuff) later.

* Setup linting

As always this was more of a hassle than it feels like it should be. I
have added details describing most of what is going on here to the
firebase folder readme.

This also includes the changes to make firestore-tests pass the linting
(just following the config rules from firebase). Doesn't include
functions changes yet because I am in the process of actually writing
those functions.

* Add a two space indent rule to eslint

For some reason it was four in _some_ places, even when as far as I can
tell the Google style that we are following should require two.

* Enforce positive in values in rules and test that

The code for running these tests is now complicated enough that I almost
certainly should have just used a test driver, but I don't think I have
the time to get back and fix that now, so unless somebody _really_ wants
it, we are going to have to just settle for my somewhat impromptu test
driver.

* Add handling of user and game watchlist syncing

I now have a firebase function that updates the watching users
collection on a game whenever a users watchlist item gets created
corresponding to that game.

I had to be a little careful because it is very easy to create infinite
loops here. I used zod out of habit, to ensure that all the unknown
objects I was injesting were okay, I hope that's alright.