Add a firestore data uploader (#39)
This is designed to upload a whole bunch of steam games from a file to
firebase. For the testing of the search, but also probably eventually
for the population of the live firebase.

It turns out we actually only have 10,000 free reads a day, so we
definitely won't have a full database by the demo. I have cut the file
to upload down to the first 1k, and added in the ones that were already
in our examples that weren't in that. Feel free to modify this data file
to add any others that you would like us to be able to search for. It
would probably be good to generally capture popular games for the demo,
but I'm sure it will be fine if we just tell them that we only have a
subset of the data.

This uses the emulator by default, but it should also be able to work on
actual firebase (I've only inserted one thing there and it was by
accident). You will need to get a firebase service key, but instructions
for all that are in the readme.