Add Shashway's initial game details page work (#40)
This message is written by James even though the commit is attributed to
(correctly, as the coding work has been done by) Shashwat.

I gave Shashway instructions for fixing a git issue that resulted in the
PR he made having a lot of changes it shouldn't have, essentially
reverting large swathes of the codebase to an earlier state. I have
hence made a new branch and copied only his relevant changes onto it.

Essentially, this:
- Adds game removal functionality.
- Adds threshold editing functionality.
- Adds the game details page on image clicks.

This crashes when you attempt to get to the game details page from the
search page, because it attempts an unsafe cast from game to watchlist
game, when the games are actually search result games. I haven't changed
this because it is in the original code that Shashwat wrote. However, I
have spoken with Shashwat and we have agreed that, with the time we have
left, it is best to just not allow the user to access the game details
page from the search results. They'll just have to add the game first. I
think this makes a lot of sense considering with nobody to add the
direct fetch of data from the steam API, all we could put on that
details page would be the name.

Liza (because it relates to search) or I should be able to fix this
later, but I think I want to include the bug in this merge for
simplicity, and so as to not complicate authorship of commits.

The only changes I have made to the code were to follow Android Studio
prompts regarding things such as variable privacy, not including
newlines, and adhering to style guidlines such as the placement of open
braces and line length. Otherwise it is exactly what Shashwat wrote.

Co-authored-by: shashwatgupta2000 <gupshashwat@gmail.com>