Update firebase get steam data (#41)
* Added a function to get steam data for a game using appid, and a function to run it every 300 minutes

* Get automatic price updates working

I have built off the work of Fahkruddin in this branch, but as he had
not been able to run the code there were a lot of necessary changes.

As it now exists, every two hours a firebase function will run that
attempts to populate the price data of every game we have tracked on
firebase. It does this all in a single http query to the steam API.
Eventually, with enough games, that might become unsustainable because
of the amount of data returned, but for now it should be fine.

I have also kept, fixed, expanded, and abstracted the function that
Fakhruddin made to do the actual getting of the price data. This should
hopefully be easy enough to include in order code to get the price data
of a particular game when adding it to the watchlist from the search
page.

I was unable to get the pubsub schedule stuff happening on the emulator,
I have no idea why, I resorted to just testing by deploying to firebase,
which sucks, but we don't have much time left. As such, these changes
are currently deployed.

Co-authored-by: Fakhruddin Nalawala <fakhruddin@nalawala.com>