Finally get the un-deleting working
The transaction we being rolled back, I believe, though I'm still not
entirely sure why. I guess to do with the exeception that was being
thrown? I guess seek forgiveness not permission is not really acceptable
in room transactions, which I suppose makes sense because the insert did
still fail.