Remove the dark mode theme (#44)
Because we are always in dark mode.

It was just another theme file that we would have needed to keep
updated, and, indeed, hadn't been. This meant that when the phone was in
dark mode the native top bar would show back up and the colours would be
wrong. As we aren't exactly going to be implementing a light and
darkmode at this stage, I think this is reasonable.