Game details fix (#45)
* remove redundant features

* actually add the changes

* Fix a couple of issues causing crashes

All to do with casting of the game to a watchlist game in the game
adapter. I probably should have just made two games from the start...
this is clearly causing problems.

* add ui changes to the game details page

this improves usability. note that the UI is mostly non functional.

* hook up view model to the game details activity

i am reusing the watchlist view model bacause it controls the database
changes.

* remove logs

* add to string resource

Co-authored-by: James Wright <hello@jameswright.me>