Automatically add new games
When a user adds a new game to their watchlist, the document is also
added to the games collection so that we start tracking it. At the
moment games are never deleted from this collection, that is, we don't
stop tracking them if all the users stop tracking them, but I really
don't think that is worth getting done today.