Add very simple firebase "syncing"
All this does is that when the user adds a game to their watchlist, it
adds it to firebase. It does not do it using the worker, and it doesn't
ever update or remove that entry.

I have messed around with a bunch of the kotlin code for connecting to
firebase, and I think I understand it now. I have also removed some bits
that I think are unnecessary. We no longer create a user object, and we
don't require the device id for things that don't involve knowing the
device.

However, it required a lot of work to get to here, so this is probably
all I am going to do on this branch, just so it can be merged to main so
I don't diverge any more.

There is still some kind of issue preventing messages getting through.