Syncing local database with firestore (#48)
This add the barest bone definition of syncing the firestore, and it was still so much work. However, now, when you add a game to your watchlist on the app, it also gets added to your watchlist on firebase. Further, if it isn't having it's price tracked, firebase will go and get its price, and start tracking it. This should in turn mean that you get a notification if your threshold was above the cost.

It's only adding though, and it doesn't happen in a worker. That comes next, but I just needed to make a PR as soon as it was in a "technically achieved what I set out to and it isn't broken" state, so that I stopped my continual divergence from main.

I have deployed this to the live firebase so that it can be tested, so please, give it a go. Checkout this branch and add a game (it will need to be a game that is on firebase, and if you try this when I haven't yet merged main into this branch you will still have the hardcoded games, so use Portal, Portal 2, or Fallout 3, all of which I know are on the list). Make sure you get the notification if it's price is below the hard coded threshold, and make sure that it turns up in firebase.

There is probably stuff I have missed in here, but I don't really have time to write more, and the commit messages are good-ish, so feel free to check them out for more details.


* Attempt to prepare the local database for syncing

Things like only marking things as inactive rather than deleting them,
and adding the created and updated times and whatnot. This is very much
a work in progress though, as I have spent so long just trying to get
adding back a marked-as-deleted game working to no avail.

* Finally get the un-deleting working

The transaction we being rolled back, I believe, though I'm still not
entirely sure why. I guess to do with the exeception that was being
thrown? I guess seek forgiveness not permission is not really acceptable
in room transactions, which I suppose makes sense because the insert did
still fail.

* Update the firestore rules

To expect the user to provide the timestamps, and also to simplify them
all a little.

* Update functions to expect client timestamps

I have removed the timestamp setting functionality from the functions,
they just accept the user ones. I believe that I also fixed a bug in my
old code where it would write the app ID rather than the watching user
ID to the games watchlist.

This also includes a very serious re-arrangement of the firebase
functions directory, to hopefully increase abstraction and project
readability.

* Automatically add new games

When a user adds a new game to their watchlist, the document is also
added to the games collection so that we start tracking it. At the
moment games are never deleted from this collection, that is, we don't
stop tracking them if all the users stop tracking them, but I really
don't think that is worth getting done today.

* Add very simple firebase "syncing"

All this does is that when the user adds a game to their watchlist, it
adds it to firebase. It does not do it using the worker, and it doesn't
ever update or remove that entry.

I have messed around with a bunch of the kotlin code for connecting to
firebase, and I think I understand it now. I have also removed some bits
that I think are unnecessary. We no longer create a user object, and we
don't require the device id for things that don't involve knowing the
device.

However, it required a lot of work to get to here, so this is probably
all I am going to do on this branch, just so it can be merged to main so
I don't diverge any more.

There is still some kind of issue preventing messages getting through.

* Ensure that notifications still work

It was pretty cool to add a game to the watchlist and then immediately
see the notification in the log...

This required a bit of secrets wrangling and a tiny bit of zod
wrangling. Seems like I haven't seriously broken anything though, which
is awesome.

* Change the default back to not using the emulator

Because other people would probably prefer to not have to launch it for
the app to work...

* Stop trying to use a local secret in the functions

Good for when running on the emulators... but not for when deploying to
firebase.