Add final changes for firebase
This adds one way syncing, other than deletes, from the phone to
firebase. I'm not getting more than that done.

It was should add a toast and updates when the app is open and the
notification happens, but that doesn't work.

I also fix the notification message to actually format in dollars. This
change is currently deployed to firebase.