Setup toast notifications when the app is open
This took a whole bunch of rearranging... but I believe the setup makes
sense now. When a notificiation is received with the app open, a toast
pops up, and the price of the game locally is updated. If you follow a
notification into the app the price of the game locally is also updated.
If the app is closed when the notification happens, but you don't follow
the notification in, it still won't be updated, but I'm trying to work
on resolving that.

This also includes a change to the firebase function, to make the
notificaiton message short enough to fit in the toast (at least on the
phone I am testing on).