Add syncing with firebase on delete
Now if you remove an item from your watchlist locally, it will also be
removed from your watchlist on Firebase.

This is all done by marking things as not active, in line with the
changes I made on the local database ages back, though the record in the
games collection is deleted for simplicity.

If this is deployed to firebase now it will actually break existing
apps, so I am holding off on that.