Add syncing with firebase watchlist items
Whenever the watchlist activity refreshes, it now adds a job to update
the local watchlist items from firebase. This comes in two parts, an
update to prices and names, and an update to the watchlist items
themselves, because those things are different documents on firebase.

This completes the syncing with firebase. Now changes on the phone are
reflected on firebase, and changes on firebase are reflected on the
phone.

This doesn't subscribe so that I could maintain the worker-managed
network and battery efficiency garuantees.

This also includes some required firebase rules updates, and a fix to
the user rules, which were broken for reading.