Various firebase fixes (#53)
* Fix firebase game deletion bug

Whoops...

When a user watch list record was deleted, it deleted the game rather
than deleting the watcher record on the game.

* Setup toast notifications when the app is open

This took a whole bunch of rearranging... but I believe the setup makes
sense now. When a notificiation is received with the app open, a toast
pops up, and the price of the game locally is updated. If you follow a
notification into the app the price of the game locally is also updated.
If the app is closed when the notification happens, but you don't follow
the notification in, it still won't be updated, but I'm trying to work
on resolving that.

This also includes a change to the firebase function, to make the
notificaiton message short enough to fit in the toast (at least on the
phone I am testing on).

* Add syncing with firebase on delete

Now if you remove an item from your watchlist locally, it will also be
removed from your watchlist on Firebase.

This is all done by marking things as not active, in line with the
changes I made on the local database ages back, though the record in the
games collection is deleted for simplicity.

If this is deployed to firebase now it will actually break existing
apps, so I am holding off on that.

* Add syncing with firebase watchlist items

Whenever the watchlist activity refreshes, it now adds a job to update
the local watchlist items from firebase. This comes in two parts, an
update to prices and names, and an update to the watchlist items
themselves, because those things are different documents on firebase.

This completes the syncing with firebase. Now changes on the phone are
reflected on firebase, and changes on firebase are reflected on the
phone.

This doesn't subscribe so that I could maintain the worker-managed
network and battery efficiency garuantees.

This also includes some required firebase rules updates, and a fix to
the user rules, which were broken for reading.