Re-comment functions credential
This isn't, like, a leak of a secret or anything, just recommenting the
line that includes the already very ignored functions credential file.

I had left it uncommented for when I was testing using the emulator,
because then it is needed to use FCM, but it doesn't work on the
deployed functions, so I have re-commented it and re-deployed.

It would probably be better to have a solution that automatically
detected if it was on the emulator... but now probably isn't the time
for that.